+++ scripts/api.js
@@ -4,6 +4,9 @@ import { fullscreenCrit } from "./animations/crits/fullscreenCrit.js";
 import { persona5Crit } from "./animations/crits/persona5Crit.js";
 import { demonSlayerTransition } from "./animations/demon-slayer-transition.js";
 import { endCredits } from "./animations/endCredits.js";
+import { eldenRingDeath } from "./animations/fromSoftwareText/deathEldenRing.js";
+import { sekiroDeath } from "./animations/fromSoftwareText/deathSekiro.js";
+import { eldenRingNounVerbed } from "./animations/fromSoftwareText/nounVerbedEldenRing.js";
 import { jojoMenacingScreenSize } from "./animations/jojo-menacing-screen-size.js";
 import { DEFAULT_CRIT_CONFIG } from "./lib/const.js";
 
@@ -38,6 +41,15 @@ export function setupAPI() {
             users || [...game.users.keys()],
             config || DEFAULT_CRIT_CONFIG
           ),
+      },
+      fromSoftware: {
+        death: {
+          eldenRing: eldenRingDeath,
+          sekiro: sekiroDeath,
+        },
+        nounVerbed: {
+          eldenRing: eldenRingNounVerbed,
+        },
       },
     },
   };